[fixup][table-api-bridge] Extract common code for java and scala StreamTableEnvironmentImpl classes
